# apps/api/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "TRIMME API",
    "description": "Public, customer, shop and admin API for the TRIMME salon and barber booking marketplace.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/media/{mediaId}": {
      "get": {
        "tags": [
          "Media"
        ],
        "summary": "A stored image (JPEG, PNG or WebP). Public and immutable: a new upload always gets a new id.",
        "operationId": "GetMedia",
        "parameters": [
          {
            "name": "mediaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "304": {
            "description": "Not Modified"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}